requirements.txt
anthropic
python-dotenv==1.0.1
pydantic==2.9.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
sse-starlette==2.1.3